V1.0
Here is the V1.0 For only get result from specific Vpp and it runs on
VS2010. However it did not support the Image File stream. For the Next
Version, I will try to make it can get any result from a struct defined
by user. The input string must be same as the Vpp output definition.
P.S. The VisionPro version is 8.2